feat: effect component support 3d geometry

+++ packages/effects-webgl/src/gl-geometry.ts
@@ -425,7 +425,7 @@ export class GLGeometry extends Geometry {
         },
       },
       indices: {
-        data: new Uint16Array(fullGeometryData.indices),
+        data: new Uint32Array(fullGeometryData.indices),
       },
       drawCount: fullGeometryData.indices.length,
     };
